Cast dimer charges and spins before featurizing monomers
forward_monomer validates its inputs with

    if not torch.is_floating_point(values) or not torch.isfinite(values).all()

so an integer tensor is rejected with "total_charge must contain finite
floating values" even when every value is finite. The AP3 fused dimer LMDB
stores total_charge_A/B as int32, and forward_dimer passed them through
untouched, so the whole pair featurization pass died on its first batch with
a message pointing at data corruption rather than a dtype mismatch.

The atomic call path in qcmlforge-exp already casts with .float(); do the
same in forward_dimer. Charges are canonically integral, so the cast belongs
at the boundary rather than in the guard.

The existing forward_dimer coverage built its batch with float charges,
which is why this never surfaced. Add a regression test using the int32
dtype the datasets actually produce; it fails without the cast.

+++ src/apnet_pt/mace/encoder.py
@@ -746,18 +746,22 @@ class MACEPolarFeaturizer(torch.nn.Module):
     def forward_dimer(self, batch: Any):
         """Run A and B as separate graph batches with shared frozen weights."""
 
+        # Charges and spins are stored as integers by the dimer datasets
+        # (total_charge_A/B are int32), while forward_monomer requires floating
+        # point tensors. Cast here so an integral charge is not reported as a
+        # non-finite value.
         features_a, direct_a = self.forward_monomer(
             batch.RA,
             batch.ZA,
-            batch.total_charge_A,
-            batch.total_spin_A,
+            batch.total_charge_A.float(),
+            batch.total_spin_A.float(),
             batch=batch.molecule_ind_A,
         )
         features_b, direct_b = self.forward_monomer(
             batch.RB,
             batch.ZB,
-            batch.total_charge_B,
-            batch.total_spin_B,
+            batch.total_charge_B.float(),
+            batch.total_spin_B.float(),
             batch=batch.molecule_ind_B,
         )
         return features_a, direct_a, features_b, direct_b
